Cambio de nombre y reorganizacion de componentes
