+++ src/component/Tasks.js
@@ -1,9 +1,8 @@
 import {React, useState, useEffect} from "react";
-import DataService from "./DataService";
-import List from "./List";
+import DataService from "../service/TaskService";
+import TaskList from "./TaskList";
 
-function Logic() {
-
+export default function Tasks() {
     const testData =
         {
           "id": 4,
@@ -21,31 +20,28 @@ function Logic() {
              .then((res)=>{
                 setTaskList(res)
         })
-        }
+        },[]
     )
 
     const Handler = ()=>{
         const testVariable = [...taskList, testData]
-        console.log("Test", testVariable)
         setTaskList(testVariable)
     }
 
 
     return(
-    <div className="App">
-        <header className="App-header">
-        React To-Do List 
-        </header>
-        <form>
-        <input></input>
-        <input></input>
-        <input></input>
-        </form>
-        <button onClick={Handler}
-        >Add Task</button>
-        <List list={taskList}></List>
-    </div>
+        <div className="App">
+            <header className="App-header">
+            React To-Do List 
+            </header>
+            <form>
+            <input></input>
+            <input></input>
+            <input></input>
+            </form>
+            <button onClick={Handler}
+            >Add Task</button>
+            <TaskList list={taskList}></TaskList>
+        </div>
     )
 }
-
-export default Logic;
